rename loss dir
