+++ lib/loss_functions/cross_entropy.py
